requirement.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
pymongo
dnspython
-e .
